compelete UI

+++ ToDoList/listpushbotton.cpp
@@ -20,7 +20,12 @@ listPushBotton::listPushBotton(TaskList * tasklist, QWidget *parent)
         ui->list->setText("I Assigned");
     else
         ui->list->setText(tasklist->getName());
-    QString buttonStyle = QString("background-color: %1; color: white;").arg(tasklist->getColor());
+    QString buttonStyle = QString("QPushButton { background-color: %1; border: none; color: white;"
+                                  "padding: 10px 20px; text-align: center; text-decoration: none; "
+                                  "display: inline-block; font-size: 16px; margin: 4px 2px; cursor: pointer; "
+                                  "border-radius: 5px; }"
+                                  "QPushButton:hover { background-color: #45a049; }"
+                                  "QPushButton:pressed { background-color: #357a38; }").arg(tasklist->getColor());
     ui->list->setStyleSheet(buttonStyle);
 }
 
@@ -31,8 +36,14 @@ listPushBotton::~listPushBotton()
 
 void listPushBotton::on_deleteList_clicked()
 {
-    taskList->deleteTaskList();
-    hide();
+    QMessageBox::StandardButton reply;
+    QString tmp ="Do you want to delete this task ?!";
+    reply = QMessageBox::question(this, "confirmation", tmp,
+                                  QMessageBox::Yes | QMessageBox::No);
+    if (reply == QMessageBox::Yes) {
+        taskList->deleteTaskList();
+        hide();
+    }
 }
